requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
genanki>=0.13.0
markdown>=3.4.0
argparse>=1.4.0
json5>=0.9.0